+++ Client.UnitTests/Impl/Builder/CamundaCloudTokenProviderTest.cs
@@ -1,17 +1,13 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Net;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using NUnit.Framework;
-using Zeebe.Client.Impl.Builder;
-using Zeebe.Client.Impl.Misc;
 
-namespace Zeebe.Client
+namespace Zeebe.Client.Impl.Builder
 {
     [TestFixture]
     public class CamundaCloudTokenProviderTest
