+++ Client.UnitTests/Impl/Builder/OAuth2TokenProviderTest.cs
@@ -6,12 +6,10 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using NUnit.Framework;
-using Zeebe.Client.Impl.Builder;
 using Zeebe.Client.Impl.Misc;
 
-namespace Zeebe.Client
+namespace Zeebe.Client.Impl.Builder
 {
     [TestFixture]
     public class OAuth2TokenProviderTest
